backend/requirements.txt
Django==5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==8.4.0
djangorestframework_simplejwt
